deals_backend/requirements.txt
pandas==1.1.5
Django==3.1.4
psycopg2-binary==2.8.6
djangorestframework==3.12.2
gunicorn==20.0.4
